Fixed a CSV Parsing Error
OK, so this is a little complicated. We got a weird example CSV that wouldn't parse properly for some reason. We traced the problem to a disagreement between the frontend parser (which was a hack parser to just peek at the headers), and the backend parser. Annoyingly, the backend parser (Babyparse) had been deprecated by its parent library (Papaparse), which just added a node interface. Worse, that interface wasn't a drop-in match for the former version. So, a lot of refactoring has gone into fixing this one weird CSV. C'est la vie.